Separate the EMNIST tasks into autoencoder and digit_recognition tasks, matching the structure of other directories.
PiperOrigin-RevId: 381176979
